Rebase from the master branch
Change the monitor to adapt to the model/tasks, model/elements,etc
